Bonus gold obstacles
Gold Obstacle:

Added a rare gold obstacle with dimensions 40x40 and color Color.YELLOW.
Spawned with a 1 in 500 chance.
Unstoppable Mode:

Added a boolean isUnstoppable to track if the player is in unstoppable mode.
Used a Timer to deactivate unstoppable mode after 15 seconds.
Collision Handling:

If the player collides with a gold obstacle, activateUnstoppableMode() is called.
During unstoppable mode, collisions with regular obstacles do not decrement lives.
Visual Feedback:

Displayed "UNSTOPPABLE!" at the top of the screen during unstoppable mode.